Added strndup, simplified integer parsing.
The method of copying a string was incorrect (no nul terminator) though not an actual problem due to usage of stored string length; however, correctness is important, so I replaced the strncpy(malloc(..)...) code with a call to strndup(...).
Also, removed the integer base processing, which was added to support '0b' binary notation, which is not necessary in this implementation. Consequently, integer parsing has been simplified by using strtoll(...) to scan the integer; therefore, integer parsing now matches the C library implementation.